vmhelper/requirements.txt
discord.py == 1.3.1
multidict  == 4.3.0